requirements.txt
# Web框架
flask==3.0.0
flask-cors==4.0.0

# Neo4j
neo4j==5.15.0

# NLP
jieba==0.42.1
snownlp==0.12.3

# AI 大模型（可选）
openai>=1.0.0
anthropic>=0.18.0

# 文件解析
ebooklib==0.18
PyPDF2==3.0.1
pdfplumber==0.10.3

# 网络爬虫
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# 工具
python-dotenv==1.0.0
tqdm==4.66.1

# 测试
pytest==7.4.3
pytest-cov==4.1.0
